Avoid undefined behavior
When one runs

> ./c4 -s hello.c

the output is

1: #include <stdio.h>
2:
3: int main()
4: {
5:   printf("hello, world\n");
    ENT  0
    IMM  9433072
    PSH
    PRTF
    ADJ  1
6:   return 0;
    IMM  0
    LEV
7: }
    LEV

Now look what happens when this is interpreted by c4:

ENT 0

runs through

else if (i == ENT) { *--sp = (int)bp; bp = sp; sp = sp - *pc++; }     // enter subroutine

The problem is: When this line is executed for the first time, bp has not yet been initialized. Using the value of an uninitialized variable is undefined behavior in C - thus not permitted. When one compiles the code under Visual C++ 17, this will indeed trigger a debug point when compiled in Debug/x86 (this is how I found the bug).

This little patch fixes this undefined behavior.