Add support for LP64 targets
.. support for ILP32 targets remains.  The fate of LLP64 targets
is unknown, but they probably don't work (same as before).

This mostly consists of making 'int' and 'long' function as the
same type, mapped to 'long' internally.  This appears to be the
better route, as it maintains the relationship that sizeof(long)
== sizeof(ptr).  However, this effectively makes the emulated
environment into ILP64 on LP64.

The main cascade of changes is due to the different needs of
printf.  This also means printf("%d", 3); doesn't really work
now, since it needs to be printf("%ld", 3).  Drat.